# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger документация проекта <<В добрые руки>>",
        "description": "Open API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://in-good-hands.dev.mind4.me",
            "description": " API Server"
        }
    ],
    "paths": {
...
        "/api/admin/get_pending_posts": {
            "get": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Получение списка постов ожидающих проверки модератора",
                "operationId": "c050dc4ed0e9f62b5414224f79247a6c",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/admin/get_review_posts": {
            "get": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Получение списка постов в проверке",
                "operationId": "0633ad059fcf81251a6c82d7d1b29826",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/admin/start_checking": {
            "post": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Начать проверку поста модератором",
                "operationId": "9c55ef2c8e310ad2cb37bcf90bb0c031",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_post"
                                ],
                                "properties": {
                                    "id_post": {
                                        "type": "integer",
                                        "example": 84
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/end_checking": {
            "patch": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Завершить проверку поста модератором",
                "operationId": "e7ccc3c6a9013f94c74d80adb3c67638",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_checking",
                                    "result"
                                ],
                                "properties": {
                                    "id_checking": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "result": {
                                        "required": [
                                            "is_public",
                                            "text"
                                        ],
                                        "properties": {
                                            "is_public": {
                                                "type": "boolean"
                                            },
                                            "text": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/cancel_review": {
            "patch": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Поменять статус поста на О\\жидает проверки ",
                "operationId": "b61b56b20f358a6109ca3dd88431d710",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_checking"
                                ],
                                "properties": {
                                    "id_checking": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://in-good-hands.dev.mind4.me/oauth/authorize",
                        "tokenUrl": "https://in-good-hands.dev.mind4.me/oauth/token",
                        "refreshUrl": "https://in-good-hands.dev.mind4.me/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}